refactor: Carve out `BuildInfo`
